Create README.md
Course 5 Assets, Threats and Vulnerabilities
•	As a data owner, how do you approach securing different types of information?
•	As a data custodian, how might it be challenging to protect the privacy of other people's information?
•	How do existing and emerging technologies factor into the challenges of safe data handling?


As a data owner, I would approach securing different data types based on their classifications or sensitivity. For example, the data can be shared with public if it is classified public. If the data is classified internal, it could be shared only within the organization. If the data is classified as confidential, it could be shared with those working in the same file. If the data is classified as restricted, it is only shared with people based need to know. 

As a data custodian, in order to protect the privacy of others data, the owner should have a strong password as well as a two way of authentication depends on the data they are dealing with. And also as a custodian I should have a strong password in order to protect the customers data or asset. 

Emerging technologies such as IoT make it difficult to protect personal data. Most of these emerging technologies have not established with strong protection methods. And also all our devices including our personal information all connected which makes our sensitive information are vulnerable.